All remaining buildings
-Added all remaining buildings, still missing a few lots/bus stop positions
-Added north west/east, south west/east connections for vertex's
