+++ app/src/main/res/values/strings.xml
@@ -56,27 +56,31 @@
 
     <string name="q_lot">Q Lot</string>
     <string name="k_lot">K Lot</string>
-    <string name="student_lot">Student Parking</string>
+    <string name="i_lot">I Lot</string>
+    <string name="student_lot">Student Parking (U Lot)</string>
 
     <string name="nb_uni_at_dysart">Northbound University at Dysart</string>
     <string name="sb_uni_at_dysart">Southbound University at Dysart</string>
-    <string name="wb_dafoe_at_uni">Westbound Dafoe at University</string>
+    <string name="wb_dafoe_at_uni">Westbound Dafoe at U of M Station</string>
     <string name="eb_dafoe_at_uni">Eastbound Dafoe at University</string>
+    <string name="eb_dafoe_at_agri">Eastbound Dafoe at Agriculture</string>
     <string name="nb_maclean_uofm">Northbound Maclean at U of M </string>
-    <string name="wb_dafoe_uofm">Westbound Dafoe at U of M (School of Music)</string>
+    <string name="wb_dafoe_at_agriculture">Westbound Dafoe at Agriculture</string>
     <string name="nb_uni_chancellor">Northbound University at Chancellor Matheson</string>
 
     <string-array name="lots_bus_stops">
         <item>@string/q_lot</item>
         <item>@string/k_lot</item>
+        <item>@string/i_lot</item>
         <item>@string/student_lot</item>
         <item>@string/nb_uni_at_dysart</item>
         <item>@string/nb_uni_chancellor</item>
         <item>@string/sb_uni_at_dysart</item>
         <item>@string/wb_dafoe_at_uni</item>
         <item>@string/eb_dafoe_at_uni</item>
+        <item>@string/eb_dafoe_at_agri</item>
         <item>@string/nb_maclean_uofm</item>
-        <item>@string/nb_uni_chancellor</item>
+        <item>@string/wb_dafoe_at_agriculture</item>
         <item>@string/welcome_centre</item>
     </string-array>
